requirements.txt
streamlit
requests
pillow
#wheel
#colorama
#ipdb
#ipykernel
#yapf
#matplotlib
#pytest
#seaborn
#numpy==1.23.5
#pandas==1.5.3
#scipy==1.10.0
#scikit-learn==1.3.1
#google-cloud-storage
#db-dtypes
#pyarrow
#tqdm
#uvicorn
#h5py
#fastapi
#streamlit==1.22
#protobuf==3.19.5
## Trick to install the version of Tensorflow depending on your processor: darwin == Mac, ARM == M1
#tensorflow-macos==2.10.0; sys_platform == 'darwin' and 'ARM' in platform_version # Mac M-chips
#tensorflow==2.10.0; sys_platform == 'darwin' and 'ARM' not in platform_version # Mac Intel chips
#tensorflow==2.10.0; sys_platform != 'darwin' # Windows & Ubuntu Intel chips#
